Adding meanstack application

+++ backend/routes/posts.js
@@ -4,6 +4,7 @@ const multer = require('multer');
 const router = express.Router();
 
 const Post = require('../model/post');
+const checkAuth = require('../middleware/check-auth');
 
 const MIME_TYPE_MAP = {
   'image/png' : 'png',
@@ -27,12 +28,15 @@ const storage = multer.diskStorage({
   }
 });
 
-router.post('', multer({storage: storage}).single("image"), (req, res, next)=>{
+router.post('',
+checkAuth,
+multer({storage: storage}).single("image"), (req, res, next)=>{
   const url = req.protocol + '://' + req.get("host");
    const post = new Post({
     title: req.body.title,
     content: req.body.content,
-    imagePath: url + "/images/" + req.file.filename
+    imagePath: url + "/images/" + req.file.filename,
+    creator: req.userData.userId
    });
    post.save().then((createdPost)=> {
     res.status(201).json({
@@ -102,7 +106,9 @@ router.get('/:id', (req, res, next) => {
   })
 })
 
-router.put('/:id', multer({storage: storage}).single("image"), (req, res, next)=> {
+router.put('/:id',
+  checkAuth,
+  multer({storage: storage}).single("image"), (req, res, next)=> {
   let imagePath = req.body.imagePath;
   if(req.file){
     const url = req.protocol + '://' + req.get("host");
@@ -112,19 +118,29 @@ router.put('/:id', multer({storage: storage}).single("image"), (req, res, next)=
     _id: req.body.id,
     title: req.body.title,
     content: req.body.content,
-    imagePath: imagePath
+    imagePath: imagePath,
+    creator: req.userData.userId
    });
-   console.log(post);
-    Post.updateOne({_id: req.params.id}, post).then(results=> {
-      console.log(results);
-      res.status(200).json({message: 'Updated Successfully'})
+    Post.updateOne({_id: req.params.id, creator: req.userData.userId}, post).then(results=> {
+      if(results.nModified > 0) {
+       res.status(200).json({message: 'Updated Successfully'})
+      }else {
+       res.status(401).json({message: 'Not authorized to update the post'})
+      }
     });
 });
 
-router.delete('/:id', (req, res, next)=>{
-    Post.deleteOne({ _id: req.params.id}).then((result)=> {
+router.delete('/:id',
+    checkAuth,
+   (req, res, next)=>{
+    Post.deleteOne({ _id: req.params.id, creator: req.userData.userId }).then((result)=> {
       console.log(result);
-      res.status(200).json({message: 'Post deleted successfully'});
+      if(result.n > 0) {
+        res.status(200).json({message: 'Post deleted successfully'});
+      } else {
+        res.status(401).json({message: 'User is not authorized to delete post'});
+      }
+
     });
 });
